requirements.txt
pytimeparse~=1.1.8
firebase-admin~=6.1.0
flask~=3.0.3
requests~=2.32.0
octorest~=0.4
netifaces~=0.11.0
opencv-python~=4.9.0.80
Flask-HTTPAuth~=4.8.0
pyserial~=3.5
numpy<2
python-nmap~=0.7.1
